+++ src/k8s/pkg/client/kubernetes/client.go
@@ -1,4 +1,4 @@
-package k8s
+package kubernetes
 
 import (
 	"fmt"
