+++ src/k8s/pkg/client/kubernetes/configmap_test.go
@@ -1,4 +1,4 @@
-package k8s
+package kubernetes
 
 import (
 	"context"
