+++ src/k8s/pkg/client/kubernetes/restart_deployment.go
@@ -1,4 +1,4 @@
-package k8s
+package kubernetes
 
 import (
 	"context"
